#68: Patch vulnerability
I can't find a perfect solution. This successfully sandboxes the preview
from test of the page (window.parent.contentDocument doesn't work
anymore). To do this, I had to actually place the re-rendering JS inside
the iframe since nested iframes with different permissions aren't
supported.

The result is that document.write calls in preview are no longer allowed
as that breaks the preview code entirely since everything is in one
iframe.